Adds route calulation

+++ app/src/main/java/com/io/vastra/data/entities/RunDescription.kt
@@ -1,5 +1,6 @@
 package com.io.vastra.data.entities
 
+import com.io.vastra.running.running_view_model.RunBreakpoint
 import java.io.Serializable
 
 class RunDescription: Serializable {
@@ -7,6 +8,6 @@ class RunDescription: Serializable {
     var route: List<RoutePoint> = listOf();
     var runDuration: Long? = 0L;
     var calories: Int? = null;
-    var distance: Double = 0.0;
+    var distance: Int = 0;
     var pacePerKm: List<Double> = listOf();
 }
